third_party/blink/web_tests/fast/dom/shadow/focus-navigation-skips-non-focusable-shadow-in-iframe-expected.txt

Tests whether focus can navigate between the two buttons by pressing tab twice.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.activeElement is iframe
PASS iframeDocument.activeElement is input
PASS document.activeElement is button2
PASS document.activeElement is iframe
PASS iframeDocument.activeElement is input
PASS document.activeElement is button1
PASS successfullyParsed is true

TEST COMPLETE

  
